requirements.txt
aiohttp
configparser

docling
future

# Basic modules
gensim
neo4j
opentelemetry-api
opentelemetry-distro
opentelemetry-exporter-otlp

# Additional Packages for export Functionality
pandas>=2.0.0

# Extra libraries are installed when needed using pipmaster

pipmaster
pydantic
python-dotenv

setuptools
tenacity

# LLM packages
tiktoken
tortoise-orm
xlsxwriter>=3.1.0
